gol/pyproject.toml
[project]
name = "gol-srv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.0",
    "drf-spectacular>=0.29.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "faker>=40.11.0",
    "ruff>=0.15.7",
]
